move to singleton sequence pattern
This gives frame control back to player.c (yay), and reduces a lot of awkward pointer sharing required for fetching various bits of Sequence metadata.

The FILE pointer that backs Sequence has been moved to a named extern with a new mutex. The mutex is wired in, but currently "useless" since there isn't any threading taking advantage of it.